+++ daemon/oci_linux.go
@@ -997,9 +997,22 @@ func WithUser(c *container.Container) coci.SpecOpts {
 		if s.Process == nil {
 			s.Process = &specs.Process{}
 		}
-		var err error
-		s.Process.User, err = getUser(c, c.Config.User)
+		user, err := getUser(c, c.Config.User)
+		// Preserve fields already set on the process user, such as Umask.
+		s.Process.User.UID = user.UID
+		s.Process.User.GID = user.GID
+		s.Process.User.AdditionalGids = user.AdditionalGids
 		return err
+	}
+}
+
+// WithUmask sets the container's umask.
+func WithUmask(c *container.Container) coci.SpecOpts {
+	return func(ctx context.Context, client coci.Client, ctr *containers.Container, s *coci.Spec) error {
+		if c.HostConfig.Umask == nil {
+			return nil
+		}
+		return coci.WithUmask(*c.HostConfig.Umask)(ctx, client, ctr, s)
 	}
 }
 
@@ -1015,6 +1028,7 @@ func (daemon *Daemon) createSpec(ctx context.Context, daemonCfg *configStore, c 
 		WithSysctls(c),
 		// Set the user before CDI device injection, which may append supplementary groups.
 		WithUser(c),
+		WithUmask(c),
 		WithDevices(daemon, c),
 		withRlimits(daemon, &daemonCfg.Config, c),
 		WithNamespaces(daemon, c),